+++ src/app/SBE/detail-profile/helpers/get-detail-type.js
