+++ src/app/SBE/detail-profile/helpers/set-top-height.js
