+++ src/app/SBE/detail-profile/template/attributes-list/index.js
